Started tokenizer code, got rid of all LONG based offsets, tokenizer is working but not done.

+++ PDFParser.Parser/Document/CrossReferenceTable.cs
@@ -5,17 +5,17 @@ namespace PDFParser.Parser.Document;
 public class CrossReferenceTable : DirectObject
 {
 
-    private readonly Dictionary<IndirectReference, long> _objectOffsets;
+    private readonly Dictionary<IndirectReference, int> _objectOffsets;
 
-    public IReadOnlyDictionary<IndirectReference, long> ObjectOffsets => _objectOffsets;
+    public IReadOnlyDictionary<IndirectReference, int> ObjectOffsets => _objectOffsets;
     
     public CrossReferenceTable(
-        IReadOnlyDictionary<IndirectReference, long> objectOffsets,
+        IReadOnlyDictionary<IndirectReference, int> objectOffsets,
         long offset,
         long length
     ) : base(offset, length)
     {
-        _objectOffsets = new Dictionary<IndirectReference, long>(objectOffsets);
+        _objectOffsets = new Dictionary<IndirectReference, int>(objectOffsets);
     }
