+++ PDFParser.Parser/IO/IMatcher.cs
@@ -2,5 +2,5 @@ namespace PDFParser.Parser.IO;
 
 public interface IMatcher
 {
-     long? FindFirstOffset(ReadOnlyMemory<byte> stream, ReadOnlySpan<byte> pattern);
+     int? FindFirstOffset(ReadOnlyMemory<byte> stream, ReadOnlySpan<byte> pattern);
 }
